+++ applications/fishing-map/src/features/app/App.tsx
@@ -1,11 +1,16 @@
 import React, { useState, useCallback, useEffect, Suspense, useLayoutEffect } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
+import { useSelector } from 'react-redux'
 import SplitView from '@globalfishingwatch/ui-components/dist/split-view'
 import Menu from '@globalfishingwatch/ui-components/dist/menu'
 import Modal from '@globalfishingwatch/ui-components/dist/modal'
 import { MapContext } from 'features/map/map-context.hooks'
 import useDebugMenu from 'features/debug/debug.hooks'
-import { isWorkspaceLocation, selectLocationType, selectWorkspaceId } from 'routes/routes.selectors'
+import {
+  isWorkspaceLocation,
+  selectLocationType,
+  selectUrlViewport,
+  selectWorkspaceId,
+} from 'routes/routes.selectors'
 import menuBgImage from 'assets/images/menubg.jpg'
 import { useLocationConnect } from 'routes/routes.hook'
 import DebugMenu from 'features/debug/DebugMenu'
@@ -28,7 +33,8 @@ import { AsyncReducerStatus } from 'utils/async-slice'
 import useViewport, { useMapFitBounds } from 'features/map/map-viewport.hooks'
 import { selectIsAnalyzing } from 'features/analysis/analysis.selectors'
 import styles from './App.module.css'
-import { selectAnalysisQuery, selectSidebarOpen, selectViewport } from './app.selectors'
+import { selectAnalysisQuery, selectSidebarOpen } from './app.selectors'
+import { useAppDispatch } from './app.hooks'
 
 const Main = () => {
   const workspaceLocation = useSelector(isWorkspaceLocation)
@@ -43,7 +49,7 @@ const Main = () => {
 }
 
 function App(): React.ReactElement {
-  const dispatch = useDispatch()
+  const dispatch = useAppDispatch()
   const sidebarOpen = useSelector(selectSidebarOpen)
   const { dispatchQueryParams } = useLocationConnect()
   const [menuOpen, setMenuOpen] = useState(false)
@@ -56,7 +62,7 @@ function App(): React.ReactElement {
   const workspaceLocation = useSelector(isWorkspaceLocation)
   const isAnalysing = useSelector(selectIsAnalyzing)
   const narrowSidebar = workspaceLocation && !analysisQuery
-  const urlViewport = useSelector(selectViewport)
+  const urlViewport = useSelector(selectUrlViewport)
 
   const fitMapBounds = useMapFitBounds()
   const { setMapCoordinates } = useViewport()
@@ -68,8 +74,6 @@ function App(): React.ReactElement {
       } else {
         setMapCoordinates({ latitude: 0, longitude: 0, zoom: 0 })
       }
-    } else {
-      setMapCoordinates(urlViewport)
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
@@ -105,12 +109,21 @@ function App(): React.ReactElement {
   const homeNeedsFetch = isHomeLocation && currentWorkspaceId !== DEFAULT_WORKSPACE_ID
   const hasWorkspaceIdChanged = locationType === WORKSPACE && currentWorkspaceId !== urlWorkspaceId
   useEffect(() => {
+    const fetchWorkspace = async () => {
+      const action = await dispatch(fetchWorkspaceThunk(urlWorkspaceId as string))
+      if (fetchWorkspaceThunk.fulfilled.match(action)) {
+        const newViewport = urlViewport || action.payload?.viewport
+        if (newViewport) {
+          setMapCoordinates(newViewport)
+        }
+      }
+    }
     if (
       userLogged &&
       workspaceCustomStatus !== AsyncReducerStatus.Loading &&
       (homeNeedsFetch || hasWorkspaceIdChanged)
     ) {
-      dispatch(fetchWorkspaceThunk(urlWorkspaceId as string))
+      fetchWorkspace()
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [userLogged, homeNeedsFetch, hasWorkspaceIdChanged])