+++ applications/fishing-map/src/features/i18n/i18nNumber.tsx
@@ -1,4 +1,4 @@
-import React, { Fragment } from 'react'
+import { Fragment } from 'react'
 import { useTranslation } from 'react-i18next'
 import { Locale } from 'types'
 import i18n from './i18n'