+++ applications/fishing-map/src/features/workspace/Workspace.tsx
@@ -7,7 +7,6 @@ import GFWAPI from '@globalfishingwatch/api-client'
 import Search from 'features/search/Search'
 import {
   selectWorkspaceStatus,
-  selectDataviewsResourceQueries,
   selectWorkspaceError,
   selectWorkspace,
 } from 'features/workspace/workspace.selectors'
@@ -21,8 +20,10 @@ import { logoutUserThunk, selectUserData } from 'features/user/user.slice'
 import { selectSearchQuery } from 'features/app/app.selectors'
 import { SUPPORT_EMAIL } from 'data/config'
 import { WorkspaceCategories } from 'data/workspaces'
+import { selectDataviewsResourceQueries } from 'features/resources/resources.selectors'
 import HeatmapsSection from './heatmaps/HeatmapsSection'
 import VesselsSection from './vessels/VesselsSection'
+import EventsSection from './events/EventsSection'
 import EnvironmentalSection from './environmental/EnvironmentalSection'
 import ContextAreaSection from './context-areas/ContextAreaSection'
 import styles from './Workspace.module.css'
@@ -151,6 +152,7 @@ function Workspace() {
       )}
       <HeatmapsSection />
       <VesselsSection />
+      <EventsSection />
       <EnvironmentalSection />
       <ContextAreaSection />
     </Fragment>