+++ packages/layer-composer/src/generators/index.ts
@@ -10,6 +10,7 @@ import UserContextGenerator from './user-context/user-context'
 import TrackGenerator from './track/track'
 import VesselEventsGenerator from './vessel-events/vessel-events'
 import RulersGenerator from './rulers/rulers'
+import TileClusterGenerator from './tile-cluster/tile-cluster'
 
 export { HEATMAP_COLOR_RAMPS } from './heatmap/config'
 export { TEMPORALGRID_SOURCE_LAYER } from './heatmap/modes/gridded'
@@ -21,6 +22,7 @@ const GeneratorConfig = {
   [Type.GL]: new GLStyleGenerator(),
   [Type.CartoPolygons]: new CartoGenerator({ baseUrl: CARTO_FISHING_MAP_API }),
   [Type.Context]: new ContextGenerator(),
+  [Type.TileCluster]: new TileClusterGenerator(),
   [Type.UserContext]: new UserContextGenerator(),
   [Type.HeatmapAnimated]: new HeatmapAnimatedGenerator(),
   [Type.Heatmap]: new HeatmapGenerator(),